refactor reportInsightViewed to eventUsageLogic
